requirements.txt
Flask==3.0.2
gunicorn==22.0.0
line-bot-sdk==2.4.2
